Adding cypress - can talk to server.

+++ e2e/cypress/support/commands.ts
@@ -1,5 +1,10 @@
 import "cypress-testing-library/add-commands";
+import { MutationOptions } from "apollo-client/core/watchQueryOptions";
+import { FetchResult } from "react-apollo";
+
 import { mockWindowsFetch } from "./mock-windows-fetch";
+import buildClientCache from "../../../src/apollo-setup";
+import { getBackendUrls } from "../../../src/get-backend-urls";
 
 declare global {
   namespace Cypress {
@@ -18,30 +23,50 @@ declare global {
        *
        */
       waitForFetches: () => Chainable<Promise<void>>;
+
+      /**
+       *
+       */
+      mutate: <TData, TVariables>(
+        options: MutationOptions<TData, TVariables>
+      ) => Promise<
+        // tslint:disable-next-line: no-any
+        FetchResult<TData, Record<string, any>, Record<string, any>>
+      >;
     }
   }
 }
 
-const serverUrl = "http://localhost:4017";
+const serverUrl = getBackendUrls(Cypress.env("apiUri"));
 
 Cypress.Commands.add("checkoutSession", async () => {
-  const response = await fetch(serverUrl + "/sandbox", {
+  const response = await fetch(serverUrl.root + "/sandbox", {
     cache: "no-store",
     method: "POST"
   });
 
-  const sessionId = await response.text();
-  return Cypress.env("sessionId", sessionId);
+  if (response.ok) {
+    const sessionId = await response.text();
+    return Cypress.env("sessionId", sessionId);
+  }
+
+  return Cypress.env("sessionId", null);
 });
 
-Cypress.Commands.add("dropSession", () =>
+Cypress.Commands.add("dropSession", () => {
+  const sessionId = Cypress.env("sessionId");
+
+  if (!sessionId) {
+    return;
+  }
+
   cy.waitForFetches().then(() =>
-    fetch(serverUrl + "/sandbox", {
+    fetch(serverUrl.root + "/sandbox", {
       method: "DELETE",
-      headers: { "x-session-id": Cypress.env("sessionId") || "0" }
+      headers: { "x-session-id": sessionId }
     })
-  )
-);
+  );
+});
 
 Cypress.Commands.add("waitForFetches", () => {
   if (Cypress.env("fetchCount") === 0) {
@@ -54,3 +79,16 @@ Cypress.Commands.add("waitForFetches", () => {
 Cypress.on("window:before:load", win => {
   cy.stub(win, "fetch", mockWindowsFetch(fetch));
 });
+
+Cypress.Commands.add(
+  "mutate",
+  <TData, TVariables>(options: MutationOptions<TData, TVariables>) => {
+    const { client } = buildClientCache({
+      uri: serverUrl.apiUrl,
+      headers: {
+        "x-session-id": Cypress.env("sessionId")
+      }
+    });
+    return client.mutate<TData, TVariables>(options);
+  }
+);
